refactor: move all all files in common under their own subdirectories
and a few more minor things
