+++ packages/pieces/convertkit/src/lib/common/broadcasts/index.ts
@@ -1,42 +1,12 @@
 import { Property, Validators } from '@activepieces/pieces-framework';
-import {
-  httpClient,
-  HttpMethod,
-  HttpRequest,
-} from '@activepieces/pieces-common';
-import { Broadcast } from './models';
-import { CONVERTKIT_API_URL } from '../common/constants';
 
-export const API_ENDPOINT = 'broadcasts';
-
-export const fetchBroadcasts = async (auth: string, page: number) => {
-  const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}?page=${page}`;
-  const body = {
-    api_secret: auth,
-  };
-  const request: HttpRequest = {
-    url,
-    body,
-    method: HttpMethod.GET,
-  };
-  const response = await httpClient.sendRequest<{ broadcasts: Broadcast[] }>(
-    request
-  );
-  if (response.status !== 200) {
-    throw new Error(
-      `Failed to fetch broadcasts: ${response.status} ${response.body}`
-    );
-  }
-  return response.body.broadcasts;
-};
-
-export const broadcastId = Property.ShortText({
+export const broadcatId = Property.ShortText({
   displayName: 'Broadcast Id',
   description: 'The broadcast id',
   required: true,
 });
 
-export const page = Property.Number({
+export const broadcastPageNumber = Property.Number({
   displayName: 'Page',
   description:
     'Page number. Each page of results will contain up to 50 broadcasts.',
@@ -45,7 +15,7 @@ export const page = Property.Number({
   validators: [Validators.number, Validators.nonZero],
 });
 
-export const content = Property.ShortText({
+export const broadcastContent = Property.ShortText({
   displayName: 'Content',
   description:
     "The broadcast's email content - this can contain text and simple HTML markdown (such as h1, img or p tags)",
@@ -56,7 +26,7 @@ export const description = Property.ShortText({
   description: 'An internal description of this broadcast',
   required: false,
 });
-export const emailAddress = Property.ShortText({
+export const broadcastEmailAddress = Property.ShortText({
   displayName: 'Email Address',
   description:
     "Sending email address; leave blank to use your account's default sending email address",
