+++ packages/pieces/convertkit/src/lib/common/constants.ts
@@ -1,1 +1,10 @@
 export const CONVERTKIT_API_URL = 'https://api.convertkit.com/v3';
+
+export const BROADCASTS_API_ENDPOINT = `${CONVERTKIT_API_URL}/broadcasts`;
+export const CUSTOM_FIELDS_API_ENDPOINT = `${CONVERTKIT_API_URL}/custom_fields`;
+export const FORMS_API_ENDPOINT = `${CONVERTKIT_API_URL}/forms`;
+export const PURCHASES_API_ENDPOINT = `${CONVERTKIT_API_URL}/purchases`;
+export const SEQUENCES_API_ENDPOINT = `${CONVERTKIT_API_URL}/sequences`;
+export const SUBSCRIBERS_API_ENDPOINT = `${CONVERTKIT_API_URL}/subscribers`;
+export const TAGS_API_ENDPOINT = `${CONVERTKIT_API_URL}/tags`;
+export const WEBHOOKS_API_ENDPOINT = `${CONVERTKIT_API_URL}/automations/hook`;
